requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gunicorn>=21.2.0
pydantic>=2.0.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
openai>=1.0.0
python-docx
pdfplumber
pypdf
pdfminer.six
python-dotenv
gtts
httpx
pytest
pytest-asyncio
